requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
python-dotenv==1.0.0
SQLAlchemy==2.0.20
Werkzeug==2.3.7
bcrypt==4.0.1
PyJWT==2.8.0
